Move files and components around.
Create separate src/components and src/views directories.
Per https://stackoverflow.com/q/50865828/, the convention
(or maybe just "a convention") is to put reusable components
in src/components and routing views in src/views.

Also move routing-specific logic into src/router/index.js
and create a new Toolbar component instead of dumping the
navigation UI in the App component. Remove some duplication
in the declaration of the "Sign out" menu item, too.
