+++ src/views/Login.vue
@@ -12,8 +12,8 @@ export default {
     ui.start("#firebaseui-auth-container", {
       signInSuccessUrl: '/',
       signInOptions: [
-        firebase.auth.EmailAuthProvider.PROVIDER_ID,
         firebase.auth.GoogleAuthProvider.PROVIDER_ID,
+        firebase.auth.EmailAuthProvider.PROVIDER_ID,
       ],
     });
   }
