+++ src/views/Scoreboard.vue
