+++ src/views/Statistics.vue
